二号引擎debug

+++ src/execution/trader.py
@@ -181,18 +181,18 @@ class OKXTrader:
             logger.error(f"[Trader] 获取K线数据异常: {e}")
             return []
 
-    async def market_buy(self, size: int, reduce_only: bool = False) -> Dict:
+    async def market_buy(self, size: float, reduce_only: bool = False) -> Dict:
         payload = {
             "instId": self.symbol, "tdMode": self.td_mode, "side": "buy",
-            "ordType": "market", "sz": str(int(size))
+            "ordType": "market", "sz": str(size)
         }
         if reduce_only: payload["reduceOnly"] = True
         return await self._request("POST", "/api/v5/trade/order", payload)
 
-    async def market_sell(self, size: int, reduce_only: bool = False) -> Dict:
+    async def market_sell(self, size: float, reduce_only: bool = False) -> Dict:
         payload = {
             "instId": self.symbol, "tdMode": self.td_mode, "side": "sell",
-            "ordType": "market", "sz": str(int(size))
+            "ordType": "market", "sz": str(size)
         }
         if reduce_only: payload["reduceOnly"] = True
         return await self._request("POST", "/api/v5/trade/order", payload)
